VCV-9: refactored and added annotation status badge
