+++ src/components/annotate/task-detail/page-client.tsx
@@ -11,120 +11,25 @@ import {
   CardTitle,
 } from '@/components/ui/card';
 import {
-  annotationFormSchema,
-  AnnotationFormOutput,
-  AnnotationFormInput,
-  isAbdomenStatusEnabled,
-  isSexEnabled,
-} from './annotation-form-panel/validation/annotation-form-schema';
-import {
   useAnnotationTaskProgressQuery,
   useTaskAnnotationsQuery,
 } from '@/lib/annotate/client';
-import { Toggle } from '@/components/ui/toggle';
-import { cn } from '@/lib/utils';
-import { Textarea } from '@/components/ui/textarea';
 import { formatDate } from '@/lib/shared/utils/date';
-import {
-  ArrowLeft,
-  ArrowRight,
-  CalendarDays,
-  Info,
-  Flag,
-  Save,
-} from 'lucide-react';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { ArrowLeft, ArrowRight, CalendarDays } from 'lucide-react';
+import { useCallback, useMemo, useState } from 'react';
 import { TaskProgressBreakdown } from './annotation-form-panel/task-progress-breakdown';
 import { SpecimenMetadata } from './specimen-image-panel/specimen-metadata';
 import { SpecimenImageViewer } from './specimen-image-panel/specimen-image-viewer';
-import MorphIdSelectMenu from './annotation-form-panel/morph-id-select-menu';
-import {
-  SPECIES_MORPH_IDS,
-  SEX_MORPH_IDS,
-  ABDOMEN_STATUS_MORPH_IDS,
-} from '@/lib/entities/specimen/morph-ids';
-import { toDomId } from '@/lib/shared/utils/dom';
-import {
-  Form,
-  FormField,
-  FormControl,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
+import { AnnotationForm } from './annotation-form-panel/annotation-form';
+
 interface AnnotationTaskDetailPageClientProps {
   taskId: number;
-  onSubmit?: (values: AnnotationFormOutput) => Promise<void> | void;
-  className?: string;
 }
 
 export function AnnotationTaskDetailPageClient({
   taskId,
-  onSubmit,
-  className,
 }: AnnotationTaskDetailPageClientProps) {
   const [page, setPage] = useState(1);
-
-  const annotationForm = useForm<AnnotationFormInput>({
-    resolver: zodResolver(annotationFormSchema),
-    defaultValues: {
-      species: undefined,
-      sex: undefined,
-      abdomenStatus: undefined,
-      notes: undefined,
-      flagged: false,
-    },
-    mode: 'onSubmit',
-    reValidateMode: 'onChange',
-  });
-  const selectedSpecies = annotationForm.watch('species');
-  const selectedSex = annotationForm.watch('sex');
-  const selectedAbdomenStatus = annotationForm.watch('abdomenStatus');
-  const isFlagged = annotationForm.watch('flagged');
-
-  const sexEnabled = isSexEnabled(selectedSpecies);
-  const abdomenStatusEnabled = isAbdomenStatusEnabled(
-    selectedSpecies,
-    selectedSex,
-  );
-
-  const handleSpeciesSelect = (newSpecies?: string) => {
-    annotationForm.setValue('species', newSpecies, { shouldDirty: true });
-    if (!isSexEnabled(newSpecies)) {
-      annotationForm.setValue('sex', undefined, {
-        shouldDirty: true,
-        shouldValidate: false,
-      });
-      annotationForm.setValue('abdomenStatus', undefined, {
-        shouldDirty: true,
-        shouldValidate: false,
-      });
-      annotationForm.clearErrors(['sex', 'abdomenStatus']);
-    }
-    annotationForm.clearErrors('species');
-  };
-
-  const handleSexSelect = (newSex?: string) => {
-    annotationForm.setValue('sex', newSex, { shouldDirty: true });
-    if (!isAbdomenStatusEnabled(selectedSpecies, newSex)) {
-      annotationForm.setValue('abdomenStatus', undefined, {
-        shouldDirty: true,
-        shouldValidate: false,
-      });
-      annotationForm.clearErrors(['abdomenStatus']);
-    }
-    annotationForm.clearErrors('sex');
-  };
-
-  const handleAbdomenStatusSelect = (newAbdomenStatus?: string) => {
-    annotationForm.setValue('abdomenStatus', newAbdomenStatus, {
-      shouldDirty: true,
-    });
-
-    annotationForm.clearErrors('abdomenStatus');
-  };
 
   const {
     data: annotationsPage,
@@ -144,6 +49,7 @@ export function AnnotationTaskDetailPageClient({
     () => annotationsPage?.items?.[0] ?? null,
     [annotationsPage],
   );
+  const annotationStatus = currentAnnotation?.status ?? null;
 
   const imageUrl = useMemo(() => {
     const specimen = currentAnnotation?.specimen;
@@ -157,40 +63,17 @@ export function AnnotationTaskDetailPageClient({
     return `/api/bff${relativePath.startsWith('/') ? relativePath : `/${relativePath}`}`;
   }, [currentAnnotation]);
 
-  const handleFlagged = (
-    newFlagged: boolean,
-    updateFormValue: (value: boolean) => void,
-  ) => {
-    updateFormValue(newFlagged);
-
-    if (newFlagged) {
-      annotationForm.clearErrors(['species', 'sex', 'abdomenStatus']);
-    } else {
-      annotationForm.clearErrors(['notes']);
-    }
-  };
-
   const handleNext = useCallback(() => {
     if (hasMore && !isFetching && !isLoading) {
       setPage(prev => prev + 1);
-      annotationForm.reset();
     }
   }, [hasMore, isFetching, isLoading]);
 
   const handlePrevious = useCallback(() => {
     if (page > 1 && !isFetching && !isLoading) {
       setPage(prev => Math.max(prev - 1, 1));
-      annotationForm.reset();
     }
   }, [page, isFetching, isLoading]);
-
-  const handleValidSubmit = (formInput: AnnotationFormInput) => {
-    const validatedFormOutput = annotationFormSchema.parse(formInput);
-    if (typeof onSubmit === 'function') {
-      return onSubmit(validatedFormOutput);
-    }
-    console.warn('onSubmit prop not provided - Data:', validatedFormOutput);
-  };
 
   const createdAt = currentAnnotation?.createdAt
     ? formatDate(currentAnnotation.createdAt).monthYear
@@ -213,6 +96,11 @@ export function AnnotationTaskDetailPageClient({
           <div className="flex flex-wrap items-center justify-between gap-3">
             <CardTitle className="text-lg font-semibold">
               Specimen Image {totalImages ? `(${page} of ${totalImages})` : ''}
+              {annotationStatus && (
+                <Badge variant="outline" className="ml-5">
+                  {annotationStatus}
+                </Badge>
+              )}
             </CardTitle>
             {createdAt && (
               <Badge
@@ -246,192 +134,31 @@ export function AnnotationTaskDetailPageClient({
           </CardTitle>
           <TaskProgressBreakdown taskProgress={taskProgress} />
         </CardHeader>
-        <div>
-          <CardContent className="pt-0">
-            <Form {...annotationForm}>
-              <form
-                onSubmit={annotationForm.handleSubmit(handleValidSubmit)}
-                className={cn('space-y-3', className)}
-              >
-                <fieldset
-                  disabled={annotationForm.formState.isSubmitting}
-                  className="space-y-3"
-                >
-                  <FormField
-                    control={annotationForm.control}
-                    name="species"
-                    render={({ field, fieldState }) => (
-                      <FormItem>
-                        <FormLabel
-                          htmlFor={toDomId(field.name)}
-                          className="m-0"
-                        >
-                          Species
-                        </FormLabel>
-                        <FormControl>
-                          <MorphIdSelectMenu
-                            label="species"
-                            morphIds={Object.values(SPECIES_MORPH_IDS)}
-                            selectedMorphId={field.value}
-                            onMorphSelect={handleSpeciesSelect}
-                            inValid={!!fieldState.error}
-                          />
-                        </FormControl>
-                        <FormMessage className="text-xs" />
-                      </FormItem>
-                    )}
-                  />
 
-                  <FormField
-                    control={annotationForm.control}
-                    name="sex"
-                    render={({ field, fieldState }) => (
-                      <FormItem>
-                        <FormLabel
-                          htmlFor={toDomId(field.name)}
-                          className="m-0"
-                        >
-                          Sex
-                        </FormLabel>
-                        <FormControl>
-                          <MorphIdSelectMenu
-                            label="Sex"
-                            morphIds={Object.values(SEX_MORPH_IDS)}
-                            selectedMorphId={field.value}
-                            onMorphSelect={handleSexSelect}
-                            inValid={!!fieldState.error && sexEnabled}
-                            disabled={!sexEnabled}
-                          />
-                        </FormControl>
-                        <FormMessage className="text-xs" />
-                      </FormItem>
-                    )}
-                  />
+        <CardContent className="pt-0">
+          <AnnotationForm key={`annotation-form-${page}`} />
+        </CardContent>
 
-                  <FormField
-                    control={annotationForm.control}
-                    name="abdomenStatus"
-                    render={({ field, fieldState }) => (
-                      <FormItem>
-                        <FormLabel
-                          htmlFor={toDomId(field.name)}
-                          className="m-0"
-                        >
-                          Abdomen Status
-                        </FormLabel>
-                        <FormControl>
-                          <MorphIdSelectMenu
-                            label="Abdomen Status"
-                            morphIds={Object.values(ABDOMEN_STATUS_MORPH_IDS)}
-                            selectedMorphId={field.value}
-                            onMorphSelect={handleAbdomenStatusSelect}
-                            inValid={!!fieldState.error && abdomenStatusEnabled}
-                            disabled={!abdomenStatusEnabled}
-                          />
-                        </FormControl>
-                        <FormMessage className="text-xs" />
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={annotationForm.control}
-                    name="notes"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel
-                          htmlFor={toDomId(field.name)}
-                          className="m-0"
-                        >
-                          Notes
-                          {isFlagged && (
-                            <span className="text-destructive">
-                              (Required)*
-                            </span>
-                          )}
-                        </FormLabel>
-                        <FormControl>
-                          <Textarea
-                            id={toDomId(field.name)}
-                            className="h-[70px] resize-none overflow-y-auto text-sm"
-                            placeholder="Add observations..."
-                            {...field}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <div className="flex gap-3 pt-2">
-                    <FormField
-                      control={annotationForm.control}
-                      name="flagged"
-                      render={({ field }) => (
-                        <Toggle
-                          pressed={field.value}
-                          onPressedChange={newFlagged =>
-                            handleFlagged(newFlagged, field.onChange)
-                          }
-                          variant="outline"
-                          disabled={annotationForm.formState.isSubmitting}
-                          className={cn(
-                            'flex-1',
-                            'flex items-center justify-center gap-1.5 rounded-md border transition-colors',
-                            'data-[state=on]:bg-destructive/10 data-[state=on]:text-destructive data-[state=on]:border-destructive data-[state=on]:hover:bg-destructive/20 motion-safe:data-[state=on]:animate-pulse',
-                            'data-[state=off]:bg-background data-[state=off]:border-input data-[state=off]:hover:bg-accent data-[state=off]:hover:text-accent-foreground',
-                          )}
-                        >
-                          <Flag
-                            className={cn(
-                              'h-4 w-4',
-                              field.value
-                                ? 'text-destructive'
-                                : 'text-muted-foreground',
-                            )}
-                          />
-                          {field.value ? 'Specimen Flagged' : 'Flag Specimen'}
-                        </Toggle>
-                      )}
-                    />
-
-                    <Button
-                      type="submit"
-                      className="flex flex-1 items-center justify-center gap-1.5"
-                      disabled={annotationForm.formState.isSubmitting}
-                    >
-                      <Save className="h-4 w-4" />
-                      {annotationForm.formState.isSubmitting
-                        ? 'Submitting...'
-                        : 'Submit'}
-                    </Button>
-                  </div>
-
-                  <CardFooter className="mt-auto flex flex-wrap items-center justify-between gap-3 px-6 py-4">
-                    <Button
-                      type="button"
-                      variant="outline"
-                      className="min-w-[120px]"
-                      onClick={handlePrevious}
-                      disabled={page === 1 || isFetching || isLoading}
-                    >
-                      <ArrowLeft className="mr-2 h-4 w-4" /> Previous
-                    </Button>
-                    <Button
-                      type="button"
-                      variant="outline"
-                      className="min-w-[120px]"
-                      onClick={handleNext}
-                      disabled={!hasMore || isFetching || isLoading}
-                    >
-                      Next <ArrowRight className="ml-2 h-4 w-4" />
-                    </Button>
-                  </CardFooter>
-                </fieldset>
-              </form>
-            </Form>
-          </CardContent>
-        </div>
+        <CardFooter className="mt-auto flex flex-wrap items-center justify-between gap-3 px-6 py-4">
+          <Button
+            type="button"
+            variant="outline"
+            className="min-w-[120px]"
+            onClick={handlePrevious}
+            disabled={page === 1 || isFetching || isLoading}
+          >
+            <ArrowLeft className="mr-2 h-4 w-4" /> Previous
+          </Button>
+          <Button
+            type="button"
+            variant="outline"
+            className="min-w-[120px]"
+            onClick={handleNext}
+            disabled={!hasMore || isFetching || isLoading}
+          >
+            Next <ArrowRight className="ml-2 h-4 w-4" />
+          </Button>
+        </CardFooter>
       </Card>
     </div>
   );
